Presion and Posicion also solved type errors in MedicionBase, where time was defined as an int instead of a float type

+++ ej1/source/MedicionBase.cpp
@@ -2,7 +2,7 @@
 #include <memory>
 using namespace std;
 
-MedicionBase::MedicionBase(int t) : tiempoMedicion(make_unique<float>(t)) {};
+MedicionBase::MedicionBase(float t) : tiempoMedicion(make_unique<float>(t)) {};
 
 float MedicionBase::getTiempo(){
   return *tiempoMedicion;
